feat:  add last_edited_time field.

+++ learning_logs/views.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 from .models import Topic, Entry
 from .forms import TopicForm, EntryForm
 from django.contrib.auth.decorators import login_required
-
+from django.utils import timezone
 
 # Create your views here.
 def index(request):
@@ -12,21 +12,23 @@ def index(request):
 
 @login_required
 def topics(request):
-    topics = Topic.objects.all()
+    topics = Topic.objects.filter(owner=request.user).all()
     context = {'topics' : topics}
     return render(request, 'learning_logs/topics.html', context)
 
 @login_required
 def topic(request, topic_id):
     topic = Topic.objects.get(id=topic_id)
-    entries = topic.entry_set.order_by('-date_added')
+    if topic.owner == request.user:
+        entries = topic.entry_set.order_by('-date_added')
 
-    context = {
-        'topic' : topic,
-        'entries' : entries
-    }
+        context = {
+            'topic' : topic,
+            'entries' : entries
+        }
 
-    return render(request, 'learning_logs/topic.html', context)
+        return render(request, 'learning_logs/topic.html', context)
+    return render(request, 'learning_logs/HTTP_404.html')
 
 @login_required
 def new_topic(request):
@@ -34,7 +36,9 @@ def new_topic(request):
     if request.method == 'POST':
         form = TopicForm(data=request.POST)
         if form.is_valid():
-            form.save()
+            new_topic = form.save(commit=False)
+            new_topic.owner = request.user
+            new_topic.save()
             return HttpResponseRedirect(reverse('topics'))
     else:
         form = TopicForm()
@@ -46,6 +50,8 @@ def new_topic(request):
 @login_required
 def delete_topic(request, topic_id):
     topic = Topic.objects.get(id=topic_id)
+    if topic.owner != request.user:
+        return render(request, 'learning_logs/HTTP_404.html')
     topic.delete()
 
     return HttpResponseRedirect(reverse('topics'))
@@ -53,7 +59,8 @@ def delete_topic(request, topic_id):
 @login_required
 def new_entry(request, topic_id):
     topic = Topic.objects.get(id=topic_id)
-
+    if topic.owner != request.user:
+        return render(request, 'learning_logs/HTTP_404.html')
     if request.method == 'POST':
         form = EntryForm(data=request.POST)
         if form.is_valid():
@@ -72,6 +79,8 @@ def new_entry(request, topic_id):
 def edit_entry(request, entry_id):
     entry = Entry.objects.get(id=entry_id)
     topic = entry.topic
+    if topic.owner != request.user:
+        return render(request, 'learning_logs/HTTP_404.html')
     if request.method == 'POST':
         form = EntryForm(instance=entry, data=request.POST)
         if form.is_valid():
@@ -89,5 +98,7 @@ def edit_entry(request, entry_id):
 def delete_entry(request, entry_id):
     entry = Entry.objects.get(id=entry_id)
     topic = entry.topic
+    if topic.owner != request.user:
+        return render(request, 'learning_logs/HTTP_404.html')
     entry.delete()
     return HttpResponseRedirect(reverse('topic', args=[topic.id]))
